fix(ADA-1736) - Add tooltip and tooltip strict position (#76)
This solves https://kaltura.atlassian.net/browse/ADA-1736. 
I've added the missing tooltip.
It sets the Tooltip position to always be bottom-left.


### Description of the Changes

Please add a detailed description of the change, whether it's an
enhancement or a bugfix.
If the PR is related to an open issue please link to it.

### CheckLists

- [ ] changes have been done against master branch, and PR does not
conflict
- [ ] new unit / functional tests have been added (whenever applicable)
- [ ] test are passing in local environment
- [ ] Travis tests are passing (or test results are not worse than on
master branch :))
- [ ] Docs have been updated